[FSDP][state_dict][dtensor][bugfix] Fix dtensor not picked up in load
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109651

Fixes https://github.com/pytorch/pytorch/issues/109648. When loading
with DTensor + device_mesh, it is not picked up that dtensor should be used.
Probably unittests don't catch this as we mostly test save + load in the same
function, and the save path sets this.

Also added a param_name to `_ext_pre_load_state_dict_transform` for better debug.
ghstack-source-id: 201260611
@exported-using-ghexport

Differential Revision: [D49435019](https://our.internmc.facebook.com/intern/diff/D49435019/)