[inductor] Fix triton compiler error in multilayer any
Fixes #109196

When we have a split reduction and the tensor is not an even multiple of the split size,
we use `ops.masked` to pad to an even multiple. In the case here we generated:
```python
tmp5 = tl.where(mask, tmp4, 0)
```

which implicitly promotes our boolean value to `int32`. The fix is to give the default
value the same dtype as `result`.

ghstack-source-id: 59266820ba8a88b2a7538a489f2e3e242ed3f252
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109325